refactor person to expert (#235)
