+++ services/base-service/spa/client/public/elements/fin-app.js
@@ -5,9 +5,9 @@ import {render} from "./fin-app.tpl.js";
 import '../elements/pages/home/app-home.js';
 import '../elements/pages/browse/app-browse.js';
 // import '../elements/pages/work/app-work.js';
-import '../elements/pages/person/app-person.js';
-import '../elements/pages/person/app-person-works-list.js';
-import '../elements/pages/person/app-person-works-list-edit.js';
+import '../elements/pages/expert/app-expert.js';
+import '../elements/pages/expert/app-expert-works-list.js';
+import '../elements/pages/expert/app-expert-works-list-edit.js';
 import '../elements/pages/search/app-search.js';
 import '../elements/pages/404/app-404.js';
 import '../elements/pages/faq/app-faq.js';
