+++ services/base-service/spa/client/public/elements/pages/expert/app-expert-works-list.tpl.js
@@ -1,0 +1,156 @@
+import { html } from 'lit';
+import { unsafeHTML } from 'lit/directives/unsafe-html.js';
+
+import { sharedStyles } from '../../styles/shared-styles';
+
+import utils from '../../../lib/utils';
+
+export function render() {
+return html`
+  <style>
+    ${sharedStyles}
+
+    :host {
+      display: block;
+    }
+
+    h1 {
+      margin-top: 0.5rem;
+      margin-bottom: 0;
+      padding-bottom: 0;
+      color: var(--color-aggie-blue);
+    }
+
+    .hero-main {
+      background: url('/images/watercolor-sage-solid.jpg') no-repeat center center;
+      background-size: 100% auto;
+      background-color: #F2FAF6;
+      width: 100%;
+      height: 12.25rem;
+    }
+
+    .hero-text {
+      padding: 2.625rem 2.625rem 4.1875rem 2.625rem
+    }
+
+    .hero-text .works {
+      display: flex;
+      align-items: center;
+    }
+
+    .hero-text .works ucdlib-icon {
+      width: 1rem;
+      height: 1rem;
+      fill: var(--color-aggie-gold);
+    }
+
+    .works span {
+      color: var(--color-black-60);
+      padding-left: 0.5rem;
+      font-size: 1rem;
+      font-weight: bold;
+      text-transform: uppercase;
+    }
+
+    .return-to-profile {
+      display: flex;
+      align-items: center;
+      cursor: pointer;
+    }
+
+    .return-to-profile ucdlib-icon {
+      fill: var(--color-aggie-gold);
+    }
+
+    .return-to-profile span {
+      padding-left: 0.5rem;
+      font-weight: bold;
+      color: var(--color-aggie-blue-80);
+    }
+
+    .main-content {
+      width: 53.5rem;
+      margin: 0 auto;
+      padding-top: 2.38rem;
+    }
+
+    .csl-bib-body, .csl-entry {
+      display: inline;
+      line-height: var(--lh-html);
+    }
+
+    .main-content .work h5 {
+      color: var(--color-aggie-blue-80);
+      margin: 0.5rem 0;
+    }
+
+    .work-details .dot {
+      padding: 0 0.25rem;
+      color: var(--black, #000);
+      font-family: Proxima Nova;
+      font-size: 1.1875rem;
+      font-style: normal;
+      font-weight: 700;
+      line-height: 1.92125rem;
+      text-transform: uppercase;
+      position: relative;
+      bottom: 0.25rem;
+    }
+
+    ucd-theme-pagination {
+      padding-bottom: 1rem;
+    }
+
+    @media (max-width: 992px) {
+      .main-content {
+        width: 90%;
+      }
+    }
+
+  </style>
+
+  <div class="content">
+    <div class="hero-main site-frame">
+      <div class="hero-text">
+        <div class="works">
+        <ucdlib-icon icon="ucdlib-experts:fa-user"></ucdlib-icon>
+        <span>${this.expertName}</span>
+        </div>
+        <h1>${this.citations.length || 0} Works</h1>
+      </div>
+    </div>
+
+    <div class="main-content">
+      <div class="return-to-profile" @click="${this._returnToProfile}">
+        <ucdlib-icon icon="ucdlib-experts:fa-circle-chevron-left"></ucdlib-icon>
+        <span>RETURN TO PROFILE</span>
+      </div>
+
+
+      ${this.citationsDisplayed.map(
+      (cite) => html`
+        <h3 style="margin: 1.19rem 0;">${cite.issued?.[0]}</h3>
+        <div class="work">
+          <h5>${unsafeHTML(cite.title)}</h5>
+          <div class="work-details">
+            <span style="min-width: fit-content;">${utils.getCitationType(cite.type)}</span>
+            <span class="dot">.</span>
+            ${unsafeHTML(cite.apa.replace('(n.d.). ', ''))}
+          </div>
+        </div>
+        <br>
+      `
+      )}
+
+      <ucd-theme-pagination
+        ?hidden="${this.paginationTotal < 2}"
+        current-page=${this.currentPage}
+        max-pages=${this.paginationTotal}
+        @page-change=${this._onPaginationChange}
+        xs-screen>
+      </ucd-theme-pagination>
+    </div>
+
+
+  </div>
+`;}
