+++ services/base-service/spa/config.js
@@ -14,7 +14,7 @@ config.client = {
   description : 'Aggie Experts is a research networking and expertise discovery tool for UC Davis.',
   appName : process.env.FIN_APP_NAME || 'spa',
   assets : (env === 'prod') ? 'dist' : 'public',
-  appRoutes : ['home', '404', 'faq', 'termsofuse', 'person', 'search', 'works', 'works-edit', 'browse'],
+  appRoutes : ['home', '404', 'faq', 'termsofuse', 'expert', 'search', 'works', 'works-edit', 'browse'],
   versions : {
     bundle : clientPackageVersion,
     loader : clientPackage.dependencies['@ucd-lib/cork-app-load'].replace(/^\D/, '')
